Use the dir and output dir that next is given
If you parse arguments into the `next export` command it would be good
if those arguments were used by `next-offline` rather than getting you
to set them again.

This reads in the `dir` and `outputDir` in the same way as the
`next-export`. Previously if you used a custom `dir` this plugin
wouldn't have worked as there was no way of telling it where your
application directory was so when it tried to locate the `BUILD_ID` it
would all break.